+++ src/procutils.h
@@ -14,4 +14,5 @@ bool is64BitProcess(HANDLE hProcess);
 HANDLE openProcess(DWORD pid, DWORD desiredAccess = PROCESS_QUERY_LIMITED_INFORMATION);
 bool getExportRva32(LPCWSTR path32, LPCSTR functionName, DWORD* pRva);
 bool getRemoteAddress32(DWORD pid, LPCSTR functionName, LPCSTR moduleName, DWORD* pAddr32);
+bool isCurrentProcessAdmin();
 }  // namespace ProcUtils